requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector storage and similarity search
faiss-cpu>=1.7.4
chromadb>=0.4.0

# LLM integrations
openai>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0

# Data processing
pydantic>=2.0.0
python-dotenv>=1.0.0

# API and web framework (optional)
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0